// src/Moryx.Orders.Web/app/src/app/api/models/recipe-model.ts
/*
 * Copyright (c) 2026 Phoenix Contact GmbH & Co. KG
 * Licensed under the Apache License, Version 2.0
*/

/* eslint-disable */
import { Entry } from '@moryx/ngx-web-framework/entry-editor';

/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { RecipeClassificationModel } from '../models/recipe-classification-model';
import { RecipeState } from '../models/recipe-state';
import { WorkplanModel } from '../models/workplan-model';
export interface RecipeModel {
  classification?: RecipeClassificationModel;
  id?: number;
  isClone?: boolean;
  name?: string | null;
  properties?: Entry;
  revision?: number;
  state?: RecipeState;
  type?: string | null;
  workplanModel?: WorkplanModel;
}